Ostatni podpunkt z lab3 i inne

+++ wsb-jpa1/src/main/java/com/jpacourse/persistance/dao/VisitDao.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface VisitDao extends Dao<VisitEntity, Long> {
 
-    List<VisitEntity> findVisitByPatientId(final int patientId);
+    List<VisitEntity> findVisitByPatientId(final Long patientId);
 }
